separate action tests into multiple modules
